+++ ompi/mca/fcoll/dynamic_gen2/fcoll_dynamic_gen2.h
@@ -12,6 +12,8 @@
  * Copyright (c) 2008-2022 University of Houston. All rights reserved.
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -56,12 +58,12 @@ int mca_fcoll_dynamic_gen2_module_finalize (ompio_file_t *file);
 
 int mca_fcoll_dynamic_gen2_file_write_all (struct ompio_file_t *fh,
                                       const void *buf,
-                                      int count,
+                                      size_t count,
                                       struct ompi_datatype_t *datatype,
                                       ompi_status_public_t * status);
 int mca_fcoll_dynamic_gen2_file_read_all (struct ompio_file_t *fh,
                                           void *buf,
-                                          int count,
+                                          size_t count,
                                           struct ompi_datatype_t *datatype,
                                           ompi_status_public_t * status);
