+++ ompi/mca/fcoll/fcoll.h
@@ -15,6 +15,8 @@
  *                         reserved.
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -115,28 +117,28 @@ typedef int (*mca_fcoll_base_module_finalize_1_0_0_fn_t)
 typedef int (*mca_fcoll_base_module_file_read_all_fn_t)
 (struct ompio_file_t *fh,
  void *buf,
- int count,
+ size_t count,
  struct ompi_datatype_t *datatype,
  ompi_status_public_t *status);
 
 typedef int (*mca_fcoll_base_module_file_iread_all_fn_t)
 (struct ompio_file_t *fh,
  void *buf,
- int count,
+ size_t count,
  struct ompi_datatype_t *datatype,
  ompi_request_t **request);
 
 typedef int (*mca_fcoll_base_module_file_write_all_fn_t)
 (struct ompio_file_t *fh,
  const void *buf,
- int count,
+ size_t count,
  struct ompi_datatype_t *datatype,
  ompi_status_public_t *status);
 
 typedef int (*mca_fcoll_base_module_file_iwrite_all_fn_t)
 (struct ompio_file_t *fh,
  const void *buf,
- int count,
+ size_t count,
  struct ompi_datatype_t *datatype,
  ompi_request_t **request);
