+++ ompi/mca/sharedfp/lockedfile/sharedfp_lockedfile_iwrite.c
@@ -12,6 +12,8 @@
  * Copyright (c) 2013-2018 University of Houston. All rights reserved.
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -31,7 +33,7 @@
 
 int mca_sharedfp_lockedfile_iwrite(ompio_file_t *fh,
                                    const void *buf,
-                                   int count,
+                                   size_t count,
                                    ompi_datatype_t *datatype,
                                    MPI_Request * request)
 {
@@ -77,7 +79,7 @@ int mca_sharedfp_lockedfile_iwrite(ompio_file_t *fh,
 
 int mca_sharedfp_lockedfile_write_ordered_begin(ompio_file_t *fh,
                                                 const void *buf,
-                                                int count,
+                                                size_t count,
                                                 struct ompi_datatype_t *datatype)
 {
     int ret = OMPI_SUCCESS;
