+++ ompi/mca/sharedfp/lockedfile/sharedfp_lockedfile_read.c
@@ -12,6 +12,8 @@
  * Copyright (c) 2013-2018 University of Houston. All rights reserved.
  * Copyright (c) 2018      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -29,7 +31,7 @@
 #include "ompi/mca/sharedfp/base/base.h"
 
 int mca_sharedfp_lockedfile_read ( ompio_file_t *fh,
-                                   void *buf, int count, MPI_Datatype datatype, MPI_Status *status)
+                                   void *buf, size_t count, MPI_Datatype datatype, MPI_Status *status)
 {
     int ret = OMPI_SUCCESS;
     OMPI_MPI_OFFSET_TYPE offset = 0;
@@ -76,7 +78,7 @@ int mca_sharedfp_lockedfile_read ( ompio_file_t *fh,
 
 int mca_sharedfp_lockedfile_read_ordered (ompio_file_t *fh,
                                            void *buf,
-                                           int count,
+                                           size_t count,
                                            struct ompi_datatype_t *datatype,
                                            ompi_status_public_t *status)
 {
