+++ ompi/mca/sharedfp/sharedfp.h
@@ -16,6 +16,8 @@
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * Copyright (c) 2016-2017 IBM Corporation. All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -125,19 +127,19 @@ typedef int (*mca_sharedfp_base_module_get_position_fn_t)(
 typedef int (*mca_sharedfp_base_module_write_fn_t)(
         struct ompio_file_t *fh,
         const void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_status_public_t *status);
 typedef int (*mca_sharedfp_base_module_write_ordered_fn_t)(
         struct ompio_file_t *fh,
         const void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_status_public_t *status);
 typedef int (*mca_sharedfp_base_module_write_ordered_begin_fn_t)(
         struct ompio_file_t *fh,
         const void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype);
 typedef int (*mca_sharedfp_base_module_write_ordered_end_fn_t)(
         struct ompio_file_t *fh,
@@ -146,31 +148,31 @@ typedef int (*mca_sharedfp_base_module_write_ordered_end_fn_t)(
 typedef int (*mca_sharedfp_base_module_iwrite_fn_t)(
         struct ompio_file_t *fh,
         const void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_request_t ** request);
 typedef int (*mca_sharedfp_base_module_read_fn_t)(
         struct ompio_file_t *fh,
         void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_status_public_t *status);
 typedef int (*mca_sharedfp_base_module_read_ordered_fn_t)(
         struct ompio_file_t *fh,
         void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_status_public_t *status);
 typedef int (*mca_sharedfp_base_module_iread_fn_t)(
         struct ompio_file_t *fh,
         void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype,
         ompi_request_t ** request);
 typedef int (*mca_sharedfp_base_module_read_ordered_begin_fn_t)(
         struct ompio_file_t *fh,
         void *buf,
-        int count,
+        size_t count,
         struct ompi_datatype_t *datatype);
 typedef int (*mca_sharedfp_base_module_read_ordered_end_fn_t)(
         struct ompio_file_t *fh,
