+++ ompi/mca/sharedfp/sm/sharedfp_sm_iwrite.c
@@ -12,6 +12,8 @@
  * Copyright (c) 2013-2018 University of Houston. All rights reserved.
  * Copyright (c) 2015-2018 Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
+ * Copyright (c) 2024      Triad National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -30,13 +32,13 @@
 
 int mca_sharedfp_sm_iwrite(ompio_file_t *fh,
                            const void *buf,
-                           int count,
+                           size_t count,
                            ompi_datatype_t *datatype,
                            MPI_Request * request)
 {
      int ret = OMPI_SUCCESS;
      OMPI_MPI_OFFSET_TYPE offset = 0;
-     long bytesRequested = 0;
+     long long bytesRequested = 0;
      size_t numofBytes;
 
      if( NULL == fh->f_sharedfp_data){
@@ -51,7 +53,7 @@ int mca_sharedfp_sm_iwrite(ompio_file_t *fh,
 
      if ( mca_sharedfp_sm_verbose ) {
          opal_output(ompi_sharedfp_base_framework.framework_output,
-		     "sharedfp_sm_iwrite: Bytes Requested is %ld\n",bytesRequested);
+		     "sharedfp_sm_iwrite: Bytes Requested is %lld\n",bytesRequested);
      }
     /* Request the offset to write bytesRequested bytes */
      ret = mca_sharedfp_sm_request_position(fh,bytesRequested,&offset);
@@ -72,7 +74,7 @@ int mca_sharedfp_sm_iwrite(ompio_file_t *fh,
 
 int mca_sharedfp_sm_write_ordered_begin(ompio_file_t *fh,
                                         const void *buf,
-                                        int count,
+                                        size_t count,
                                         struct ompi_datatype_t *datatype)
 {
     int ret = OMPI_SUCCESS;
@@ -81,7 +83,7 @@ int mca_sharedfp_sm_write_ordered_begin(ompio_file_t *fh,
     long *buff=NULL;
     long offsetBuff;
     OMPI_MPI_OFFSET_TYPE offsetReceived = 0;
-    long bytesRequested = 0;
+    long long bytesRequested = 0;
     int recvcnt = 1, sendcnt = 1;
     size_t numofBytes;
     int i;
@@ -123,7 +125,7 @@ int mca_sharedfp_sm_write_ordered_begin(ompio_file_t *fh,
 	    bytesRequested += buff[i];
 	    if ( mca_sharedfp_sm_verbose ) {
 		opal_output(ompi_sharedfp_base_framework.framework_output,
-			    "mca_sharedfp_sm_write_ordered_begin: Bytes requested are %ld\n",
+			    "mca_sharedfp_sm_write_ordered_begin: Bytes requested are %lld\n",
 			    bytesRequested);
 	    }
         }
